--- src/main/java/com/codeup/adlister/controllers/AdDetailsServlet.java
@@ -1,24 +1,0 @@
-package com.codeup.adlister.controllers;
-
-import javax.servlet.*;
-import javax.servlet.http.*;
-import javax.servlet.annotation.*;
-import java.io.IOException;
-
-@WebServlet(name = "AdDetailsServlet", value = "/hi")
-public class AdDetailsServlet extends HttpServlet {
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        if (request.getSession().getAttribute("user") == null) {
-            response.sendRedirect("/login");
-            return;
-        }
-        request.getRequestDispatcher("/WEB-INF/ads/index.jsp");
-
-    }
-
-    @Override
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-
-    }
-}
