+++ src/main/java/com/codeup/adlister/controllers/ViewAdDetailsServlet.java
@@ -4,22 +4,17 @@ import com.codeup.adlister.dao.DaoFactory;
 import com.codeup.adlister.models.Ad;
 import com.codeup.adlister.models.User;
 
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import javax.servlet.*;
+import javax.servlet.http.*;
+import javax.servlet.annotation.*;
 import java.io.IOException;
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-@WebServlet(name = "controllers.ViewAdDetailServlet", urlPatterns = "/detail")
-public class ViewAdDetailServlet extends HttpServlet {
-
+@WebServlet(name = "AdDetailsServlet", value = "/detail")
+public class ViewAdDetailsServlet extends HttpServlet {
+    @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-
         User user = (User) request.getSession().getAttribute("user");
 
         if (user == null) {
@@ -28,13 +23,17 @@ public class ViewAdDetailServlet extends HttpServlet {
         }
 
 
-        int sentId = Integer.parseInt( request.getParameter("detail_id"));
+        int sentId = Integer.parseInt( request.getParameter("adId"));
         Ad helperAd = DaoFactory.getAdsDao().findById(sentId);
 
         List<Ad> helperList = new ArrayList<>();
         helperList.add(helperAd);
         request.setAttribute("details", helperList);
-        request.getRequestDispatcher("/WEB-INF/ads/details.jsp").forward(request, response);
+        request.getRequestDispatcher("/WEB-INF/ads/detailsAll.jsp").forward(request, response);
     }
 
+    @Override
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+
+    }
 }
